Moved MediaBrowserType to it's own class
